Chapter 10 Question 7 and Chapter 7 Question 7
These are the working programs of the chapter 7 question 7 and the chapter 10 question 7

I still need to add comments to the Chapter 7 code but the chapter 10 code has all the comments and java docs completed. I will update the code with the additional comments and java docs tomorrow. Everything works and has been tested as well.